+++ src/main/java/com/bankapp/demo/controller/UserController.java
@@ -1,28 +1,50 @@
+// Copyright (c) 2021 Razvan Balasa
 package com.bankapp.demo.controller;
 
 import com.bankapp.demo.model.User;
+import com.bankapp.demo.model.UserCredentials;
+import com.bankapp.demo.service.UserCredentialsService;
 import com.bankapp.demo.service.UserService;
+import com.bankapp.demo.utils.AuthenticationManager;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Collections;
 import java.util.List;
 
 @RestController
 @RequestMapping("/user")
 public class UserController {
+
     private UserService userService;
 
-    public UserController(UserService userService) {
+    private UserCredentialsService userCredentialsService;
+
+    private AuthenticationManager authenticationManager;
+
+    public UserController(UserService userService, UserCredentialsService userCredentialsService, AuthenticationManager authenticationManager) {
         this.userService = userService;
+        this.userCredentialsService = userCredentialsService;
+        this.authenticationManager = authenticationManager;
     }
 
     @RequestMapping(method = RequestMethod.POST, path = "/save")
-    public User adauga(@RequestBody User user) {
-        return userService.save(user);
+    public ResponseEntity<User> save(@RequestBody User user) {
+        if ("admin".equals(authenticationManager.getAuthority())) {
+            return ResponseEntity.ok(userService.save(user));
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
     }
 
     @GetMapping(path = "/get")
-    public List<User> getAll() {
-        return userService.getAll();
+    public ResponseEntity<List<User>> getAll() {
+        if ("admin".equals(authenticationManager.getAuthority())) {
+            return ResponseEntity.ok(userService.getAll());
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Collections.emptyList());
+        }
     }
 
     @GetMapping(path = "/users")
@@ -35,8 +57,17 @@ public class UserController {
         return userService.getAllSuppliers();
     }
 
+    @RequestMapping(method = RequestMethod.POST, path = "/credentials")
+    public void saveCredentials(@RequestBody UserCredentials userCredentials) {
+        if ("admin".equals(authenticationManager.getAuthority())) {
+            userCredentialsService.save(userCredentials);
+        }
+    }
+
     @DeleteMapping(path = "delete/{userId}")
     public void delete(@PathVariable Long userId) {
-        userService.delete(userId);
+        if ("admin".equals(authenticationManager.getAuthority())) {
+            userService.delete(userId);
+        }
     }
 }
